requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# LLM and AI
openai==1.3.0
anthropic==0.7.0
langchain==0.0.350
langchain-community==0.0.10

# Data processing
pandas==2.1.4
numpy==1.24.4
scipy==1.11.4
scikit-learn==1.3.2

# Optimization
optuna==3.4.0
bayesian-optimization==1.4.3

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# File handling
pyyaml==6.0.1
jinja2==3.1.2
openpyxl==3.1.2

# Web interface
streamlit==1.28.2
plotly==5.17.0
dash==2.14.2

# EDA tools integration
pexpect==4.8.0
paramiko==3.3.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Utilities
click==8.1.7
rich==13.7.0
tqdm==4.66.1
python-dotenv==1.0.0

# Security
cryptography==41.0.8
python-jose[cryptography]==3.3.0
